fix(deploy): address Pages model review findings
- Fix ensurePagesDomain to use Pages /domains API (not project.domains)
- Orchestrate verify:cloudflare with dual Bitwarden creds (Pages + Workers)
- Extract shared pages-configs and cf-bw-run helpers; dedupe path lists
- Make trigger-pages-deployments idempotent via --if-empty
- Fail setup:cloudflare-deploy when Workers Builds step fails
- Load global BW creds when migrate runs setup-live-domains
