+++ scripts/lib/pages-configs.mjs
@@ -1,0 +1,14 @@
+/** Shared Pages SSOT paths (enishu model). */
+export const PAGES_CONFIG_PATHS = [
+  "infra/pages-app.json",
+  "infra/pages-app-staging.json",
+  "infra/pages-marketing.json",
+  "infra/pages-marketing-staging.json",
+];
+
+export const LEGACY_FRONTEND_WORKERS = [
+  "roxabi-live-app",
+  "roxabi-live-app-staging",
+  "roxabi-live-marketing",
+  "roxabi-live-marketing-staging",
+];
